Move RoomRepository to application package and rename ApiWordProvider package
