+++ src/test/java/com/guinardpouard/imposteur/application/RoomServiceTests.java
@@ -2,7 +2,7 @@ package com.guinardpouard.imposteur.application;
 
 import com.guinardpouard.imposteur.domain.model.Room;
 import com.guinardpouard.imposteur.application.port.GamePublisher;
-import com.guinardpouard.imposteur.domain.repository.RoomRepository;
+import com.guinardpouard.imposteur.application.port.RoomRepository;
 import com.guinardpouard.imposteur.application.mapper.RoomUpdatedMapper;
 import com.guinardpouard.imposteur.application.event.RoomUpdatedMessage;
 import org.junit.jupiter.api.BeforeEach;
