+++ src/test/java/com/guinardpouard/imposteur/infrastructure/persistence/inmemory/InMemoryRoomRepositoryTests.java
@@ -1,7 +1,7 @@
 package com.guinardpouard.imposteur.infrastructure.persistence.inmemory;
 
 import com.guinardpouard.imposteur.domain.model.Room;
-import com.guinardpouard.imposteur.domain.repository.RoomRepository;
+import com.guinardpouard.imposteur.application.port.RoomRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
